vsim: fix the frame-2666 freeze -- (int)main_time overflowed
SimBlockDevice::BeforeEval took its cycle count as `int` and the call site
truncated: blk.BeforeEval((int)main_time). main_time is a vluint64_t advancing
once per simulated 48 MHz cycle, so at 60 Hz it passes INT_MAX (2147483647) at
frame ~2666 -- roughly 800k cycles a frame. The cast then goes negative and the
first line of the function,

    if (cycles < 2000) return;      // "wait until the computer boots"

becomes permanently true. From that moment the block device silently stops
servicing every request. Widened the parameter to uint64_t and dropped the cast.

Harness bug, not RTL; hardware is unaffected. But it invalidates any vsim result
past frame ~2666, which is worth knowing before trusting a long run.
run_tests.sh uses FRAMES=620 and never reached it, which is why it hid for so
long.

It presented as a disk read hanging forever with DRQ and INTRQ both dead and the
main CPU spinning in the boot ROM's transfer loop at $ff94. With the fix the DMA
runs on to LBA 427 instead of freezing at 418 and the CPU leaves the loop; the
game then crashes into low memory, which is a separate and later fault.

Four wrong answers preceded this, every one of them from $display: a stuck
sd_ack, a stale-bit race in the ack shift register (disproved -- flushing ack
changed nothing, instruction counts byte-identical), "the block device is idle at
the stall" (from a run that timed out before reaching the stall), and "so it must
be RTL". What found it was the VCD showing the controller waiting correctly while
ce kept ticking, and then a probe that printed nothing at all because the guard
it sat behind had already gone dead. The silence was the evidence.

run_tests.sh: all 8 rows unchanged.

+++ vsim/sim/sim_blkdevice.cpp
@@ -53,7 +53,7 @@ void SimBlockDevice::MountDisk( std::string file, int index) {
 }
 
 
-void SimBlockDevice::BeforeEval(int cycles)
+void SimBlockDevice::BeforeEval(uint64_t cycles)
 {
 //
 // switch to a new disk if current_disk is -1
@@ -152,7 +152,7 @@ void SimBlockDevice::BeforeEval(int cycles)
             int track = lba / 13;  // 13 sectors per track
             int sector = lba % 13;
             printf("FLOPPY DMA: LBA=%d (track=%d sector=%d) seek=%06X reading=%d writing=%d\n",
-                   lba, track, sector, (lba) * kBLKSZ + header_size[i], reading, writing);
+                   lba, track, sector, (int)((lba) * kBLKSZ + header_size[i]), reading, writing);
         }
         bytecnt = 0;
         *sd_buff_addr = 0;
